requirements.txt
flask==3.1.1
flask-cors==6.0.1
requests==2.31.0
beautifulsoup4==4.12.2
googlemaps==4.10.0
python-dotenv==1.0.0
lxml==4.9.3
gunicorn==21.2.0 